+++ internal/cmd/gerrit/api.go
@@ -1,6 +1,7 @@
 package gerrit
 
 import (
+	_ "embed"
 	"fmt"
 	"io"
 	"log"
@@ -11,17 +12,22 @@ import (
 	"github.com/spf13/cobra"
 )
 
-var (
-	gerritAPIMethod   string
-	gerritAPIUser     string
-	gerritAPIPassword string
-)
+//go:embed api.example
+var apiExample string
 
 func NewGerritAPICmd() *cobra.Command {
+	config := LoadConfig()
+
+	var (
+		method   string
+		user     string
+		password string
+	)
+
 	cmd := &cobra.Command{
 		Use:     "api",
 		Short:   "Gerrit's API",
-		Example: `api --auth-user Username --auth-password Password accounts/addshore`,
+		Example: apiExample,
 		Long:    `https://gerrit.wikimedia.org/r/Documentation/rest-api.html`,
 		Run: func(cmd *cobra.Command, args []string) {
 			if len(args) == 0 {
@@ -29,20 +35,25 @@ func NewGerritAPICmd() *cobra.Command {
 				os.Exit(1)
 			}
 
+			if user == "" {
+				logrus.Trace("Using username from config")
+				user = config.Username
+			}
+			if password == "" {
+				logrus.Trace("Using password from config")
+				password = config.Password
+			}
+
 			instance := "https://gerrit.wikimedia.org/r/"
 			client, _ := gerrit.NewClient(instance, nil)
-			if gerritAPIUser != "" && gerritAPIPassword != "" {
+			if user != "" && password != "" {
+				client.Authentication.SetBasicAuth(user, password)
 				logrus.Trace("Using username and password")
-				client.Authentication.SetBasicAuth(gerritAPIUser, gerritAPIPassword)
 			}
 
-			resp, err := client.Call(gerritAPIMethod, args[0], nil, nil)
+			resp, err := client.Call(method, args[0], nil, nil)
 			if err != nil {
 				logrus.Fatalf("Error making request: %s", err)
-			}
-
-			if err != nil {
-				fmt.Printf("Error: %s\n", err)
 			}
 
 			defer resp.Response.Body.Close()
@@ -57,8 +68,8 @@ func NewGerritAPICmd() *cobra.Command {
 			fmt.Print(string(b))
 		},
 	}
-	cmd.Flags().StringVarP(&gerritAPIMethod, "method", "X", "GET", "The HTTP method for the request (default \"GET\")")
-	cmd.Flags().StringVarP(&gerritAPIUser, "auth-user", "", "", "Gerrit HTTP user")
-	cmd.Flags().StringVarP(&gerritAPIPassword, "auth-password", "", "", "Gerrit HTTP password")
+	cmd.Flags().StringVarP(&method, "method", "X", "GET", "The HTTP method for the request")
+	cmd.Flags().StringVarP(&user, "auth-user", "", "", "Gerrit HTTP user")
+	cmd.Flags().StringVarP(&password, "auth-password", "", "", "Gerrit HTTP password")
 	return cmd
 }