+++ CHANGELOG.md
@@ -7,6 +7,7 @@ Each tagged release MUST have a section 2 heading starting at the time of releas
 ## v0.20.0
 
 - Added XDG standards usage for config directory location (T305150)
+- Added persistent Gerrit HTTP authentication for cli commands
 - Added helpful infomation for the user when service are created, such as `mailhog`, `graphite`, `adminer` and `phpmyadmin`
 - Added long documentation text for the `graphite`, `adminer` and `phpmyadmin` commands
 - Updated `gitlab` command to [1.25.3](https://gitlab.com/gitlab-org/cli/-/releases/v1.25.3)